// src/features/dashboard/settings/schemas/user.schema.ts
import { UserRole } from "@/enums/user.role.enum";
import { z } from "zod";

export const userSchema = z.object({
  name: z.string().min(3, "Name must be at least 3 characters"),
  email: z.string().email("Invalid email address"),
  role: z.nativeEnum(UserRole),
  bio: z.string().optional(),
  avatar: z.any().optional(),
});

export type UserFormValuesType = z.infer<typeof userSchema>;